requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==24.2.0
autogen-agentchat==0.2.36
babel==2.16.0
beautifulsoup4==4.12.3
blis==0.7.11
Brotli==1.1.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
clean-text==0.6.0
click==8.1.7
confection==0.1.5
cymem==2.0.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
emoji==1.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.16.1
FLAML==2.3.1
fsspec==2024.9.0
ftfy==6.2.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonlines==4.0.0
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.24.4
openai==1.52.0
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.8.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
ptyprocess==0.7.0
pydantic==1.10.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
records==0.6.0
regex==2024.9.11
requests==2.32.3
scikit-learn==1.0.2
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
tablib==3.6.1
tabulate==0.9.0
termcolor==2.4.0
thinc==8.1.12
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.66.5
transformers==4.26.1
typer==0.7.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
